fixed edit backend
